Change location of `QueryValidationState` type.
